Refactor convertUserMetadata.
- Tweak proposal token logging
- Move the politeia API calls into a separate file called politeia.go
- Minor refactor of the convertUserMetadata function
